requirements.txt
ImageHash==4.3.2
lmdb==1.6.2
matplotlib
natsort==8.4.0
numba
numpy
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow
pybind11==2.13.6
PyYAML==6.0.2
Requests==2.32.3
rich==14.0.0
scipy
setuptools
tb_nightly
torch
torch_scatter
torch_xla
torchvision
tqdm==4.67.1
wandb==0.19.9
xxhash==3.5.0
scikit-image
open_clip_torch
